pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sage"
version = "0.1.0"
description = "Research toolkit for multi-agent reinforcement learning experiments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langfuse>=3.9.1",
    "langgraph>=1.0.2",
    "mem0ai>=1.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-ai>=1.12.0",
    "pyyaml>=6.0.3",
    "rich>=13.7.0", # Colored logging and output
    "typer>=0.20.0",
    "whisperx>=3.7.4",
    "youtube-transcript-api>=1.2.3",
    "yt-dlp>=2025.10.22",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sage"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-watch>=4.2.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "xenon>=0.9.0",
    "hatchling>=1.18.0",
]

[project.scripts]
sage = "sage.cli.main:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "unit: Unit tests that test isolated components",
    "integration: Integration tests that test component interactions",
    "contract: Contract tests that test API boundaries",
]

[tool.coverage.run]
source = ["sage"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
